+++ lib/kitchen/directions/bake_figure.rb
@@ -32,12 +32,13 @@ module Kitchen
             #{"<span class=\'os-title\' data-type=\'title\'>#{title.children}</span>" if title}
           HTML
         if figure.has_class?('mechanism-figure')
+          scaled_class = ' scaled-down' if figure.has_class?('scaled-down')
           figure.prepend(sibling:
             <<~HTML
-              <div class="os-caption-title-container">
+              <div class="os-caption-title-container#{scaled_class}">
                 #{caption_title_content}
               </div>
-              <div class="os-caption-text-container">
+              <div class="os-caption-text-container#{scaled_class}">
                 #{caption}
               </div>
             HTML